+++ src/chart_types/xy_chart/utils/axis_utils.test.ts
@@ -30,12 +30,14 @@ import { CanvasTextBBoxCalculator } from '../../../utils/bbox/canvas_text_bbox_c
 import { SvgTextBBoxCalculator } from '../../../utils/bbox/svg_text_bbox_calculator';
 import { Position, mergePartial } from '../../../utils/common';
 import { niceTimeFormatter } from '../../../utils/data/formatters';
+import { OrdinalDomain } from '../../../utils/domain';
 import { AxisId, GroupId } from '../../../utils/ids';
 import { LIGHT_THEME } from '../../../utils/themes/light_theme';
 import { AxisStyle, TextOffset } from '../../../utils/themes/theme';
 import { XDomain, YDomain } from '../domains/types';
 import { computeAxesGeometriesSelector } from '../state/selectors/compute_axes_geometries';
 import { computeAxisTicksDimensionsSelector } from '../state/selectors/compute_axis_ticks_dimensions';
+import { getScale, SmallMultipleScales } from '../state/selectors/compute_small_multiple_scales';
 import { getAxesStylesSelector } from '../state/selectors/get_axis_styles';
 import { computeGridLinesSelector } from '../state/selectors/get_grid_lines';
 import { mergeYCustomDomainsByGroupId } from '../state/selectors/merge_y_custom_domains';
@@ -60,6 +62,8 @@ import {
 } from './axis_utils';
 import { computeXScale } from './scales';
 import { AxisSpec, DomainRange, DEFAULT_GLOBAL_ID } from './specs';
+
+const NO_ROTATION = 0;
 
 const getCustomStyle = (rotation = 0, padding = 10): AxisStyle =>
   mergePartial(LIGHT_THEME.axes, {
@@ -117,6 +121,7 @@ describe('Axis computational utils', () => {
     chartType: ChartTypes.XYAxis,
     specType: SpecTypes.Axis,
     id: 'axis_1',
+    title: 'Axis 1',
     groupId: 'group_1',
     hide: false,
     showOverlappingTicks: false,
@@ -131,6 +136,7 @@ describe('Axis computational utils', () => {
     chartType: ChartTypes.XYAxis,
     specType: SpecTypes.Axis,
     id: 'axis_2',
+    title: 'Axis 2',
     groupId: 'group_1',
     hide: false,
     showOverlappingTicks: false,
@@ -170,19 +176,6 @@ describe('Axis computational utils', () => {
     integersOnly: false,
   });
 
-  // const horizontalAxisSpecWTitle: AxisSpec = {
-  //   id: ('axis_2'),
-  //   groupId: ('group_1'),
-  //   title: 'h axis',
-  //   hide: false,
-  //   showOverlappingTicks: false,
-  //   showOverlappingLabels: false,
-  //   position: Position.Top,
-  //   style,
-  //   tickFormat: (value: any) => {
-  //     return `${value}`;
-  //   },
-  // };
   const lineSeriesSpec = MockSeriesSpec.line({
     id: 'line',
     groupId: 'group_1',
@@ -212,6 +205,27 @@ describe('Axis computational utils', () => {
     isBandScale: false,
   };
 
+  const getSmScales = (smHDomain: OrdinalDomain = [], smVDomain: OrdinalDomain = []): SmallMultipleScales => ({
+    horizontal: getScale(smHDomain, chartDim.width),
+    vertical: getScale(smVDomain, chartDim.height),
+  });
+
+  const emptySmScales = getSmScales();
+
+  const axisTitleStyles = (titleHeight: number, panelTitleHeight?: number) =>
+    mergePartial(LIGHT_THEME.axes, {
+      axisTitle: {
+        fontSize: titleHeight,
+        padding: {
+          inner: 0,
+          outer: 10,
+        },
+      },
+      axisPanelTitle: {
+        fontSize: panelTitleHeight,
+      },
+    });
+
   const { axes } = LIGHT_THEME;
 
   test('should compute axis dimensions', () => {
@@ -222,7 +236,7 @@ describe('Axis computational utils', () => {
       [yDomain],
       1,
       bboxCalculator,
-      0,
+      NO_ROTATION,
       axes,
       (v) => `${v}`,
     );
@@ -235,7 +249,7 @@ describe('Axis computational utils', () => {
       [yDomain],
       1,
       bboxCalculator,
-      0,
+      NO_ROTATION,
       axes,
       (v) => `${v}`,
       undefined,
@@ -256,7 +270,7 @@ describe('Axis computational utils', () => {
       [yDomain],
       1,
       bboxCalculator,
-      0,
+      NO_ROTATION,
       axes,
       (v) => `${v}`,
     );
@@ -279,7 +293,7 @@ describe('Axis computational utils', () => {
       [yDomain],
       1,
       bboxCalculator,
-      0,
+      NO_ROTATION,
       axes,
       (v) => `${v}`,
     );
@@ -296,7 +310,7 @@ describe('Axis computational utils', () => {
       [yDomain],
       1,
       bboxCalculator,
-      0,
+      NO_ROTATION,
       axes,
       (v) => `${v}`,
     );
@@ -313,7 +327,7 @@ describe('Axis computational utils', () => {
       [yDomain],
       1,
       bboxCalculator,
-      0,
+      NO_ROTATION,
       axes,
       (v) => `${v}`,
     );
@@ -960,8 +974,6 @@ describe('Axis computational utils', () => {
   });
 
   test('should compute axis ticks positions with title', () => {
-    const chartRotation = 0;
-
     // validate assumptions for test
     expect(verticalAxisSpec.id).toEqual(verticalAxisSpecWTitle.id);
 
@@ -976,13 +988,13 @@ describe('Axis computational utils', () => {
         leftMargin: 0,
       },
       LIGHT_THEME,
-      chartRotation,
+      NO_ROTATION,
       axisSpecs,
       axisDims,
       axesStyles,
       xDomain,
       [yDomain],
-      chartDim,
+      emptySmScales,
       1,
       false,
       (v) => `${v}`,
@@ -1008,13 +1020,13 @@ describe('Axis computational utils', () => {
         leftMargin: 0,
       },
       LIGHT_THEME,
-      chartRotation,
+      NO_ROTATION,
       axisSpecs,
       axisDims,
       axesStyles,
       xDomain,
       [yDomain],
-      chartDim,
+      emptySmScales,
       1,
       false,
       (v) => `${v}`,
@@ -1025,16 +1037,9 @@ describe('Axis computational utils', () => {
       x: 10,
     });
     expect(verticalAxisSpecWTitleGeoms?.size).toEqual({
-      width: 10,
+      width: 30,
       height: 100,
     });
-  });
-
-  const axisTitleStyles = mergePartial(LIGHT_THEME.axes.axisTitle, {
-    padding: {
-      inner: 0,
-      outer: 10,
-    },
   });
 
   test('should compute left axis position', () => {
@@ -1047,10 +1052,10 @@ describe('Axis computational utils', () => {
     const leftAxisPosition = getAxisPosition(
       chartDim,
       LIGHT_THEME.chartMargins,
-      axisTitleHeight,
-      axisTitleStyles,
+      axisTitleStyles(axisTitleHeight),
       verticalAxisSpec,
       axis1Dims,
+      emptySmScales,
       cumTopSum,
       cumBottomSum,
       cumLeftSum,
@@ -1087,10 +1092,10 @@ describe('Axis computational utils', () => {
     const rightAxisPosition = getAxisPosition(
       chartDim,
       LIGHT_THEME.chartMargins,
-      axisTitleHeight,
-      axisTitleStyles,
+      axisTitleStyles(axisTitleHeight),
       verticalAxisSpec,
       axis1Dims,
+      emptySmScales,
       cumTopSum,
       cumBottomSum,
       cumLeftSum,
@@ -1127,10 +1132,10 @@ describe('Axis computational utils', () => {
     const topAxisPosition = getAxisPosition(
       chartDim,
       LIGHT_THEME.chartMargins,
-      axisTitleHeight,
-      axisTitleStyles,
+      axisTitleStyles(axisTitleHeight),
       horizontalAxisSpec,
       axis1Dims,
+      emptySmScales,
       cumTopSum,
       cumBottomSum,
       cumLeftSum,
@@ -1168,10 +1173,10 @@ describe('Axis computational utils', () => {
     const bottomAxisPosition = getAxisPosition(
       chartDim,
       LIGHT_THEME.chartMargins,
-      axisTitleHeight,
-      axisTitleStyles,
+      axisTitleStyles(axisTitleHeight),
       horizontalAxisSpec,
       axis1Dims,
+      emptySmScales,
       cumTopSum,
       cumBottomSum,
       cumLeftSum,
@@ -1198,8 +1203,6 @@ describe('Axis computational utils', () => {
   });
 
   test('should not compute axis ticks positions if missaligned specs', () => {
-    const chartRotation = 0;
-
     const axisSpecs = [verticalAxisSpec];
     const axisStyles = new Map();
     const axisDims = new Map<AxisId, AxisTicksDimensions>();
@@ -1211,13 +1214,13 @@ describe('Axis computational utils', () => {
         leftMargin: 0,
       },
       LIGHT_THEME,
-      chartRotation,
+      NO_ROTATION,
       axisSpecs,
       axisDims,
       axisStyles,
       xDomain,
       [yDomain],
-      chartDim,
+      emptySmScales,
       1,
       false,
       (v) => `${v}`,
@@ -1284,13 +1287,13 @@ describe('Axis computational utils', () => {
           leftMargin: 0,
         },
         LIGHT_THEME,
-        0,
+        NO_ROTATION,
         invalidSpecs,
         axisDims,
         axisStyles,
         xDomain,
         [yDomain],
-        chartDim,
+        emptySmScales,
         1,
         false,
         (v) => `${v}`,
@@ -1675,13 +1678,13 @@ describe('Axis computational utils', () => {
           leftMargin: 0,
         },
         LIGHT_THEME,
-        0,
+        NO_ROTATION,
         axisSpecs,
         axisDims,
         axesStyles,
         xDomain,
         [yDomain],
-        chartDim,
+        emptySmScales,
         1,
         false,
         customFormatter,
@@ -1705,13 +1708,13 @@ describe('Axis computational utils', () => {
           leftMargin: 0,
         },
         LIGHT_THEME,
-        0,
+        NO_ROTATION,
         axisSpecs,
         axisDims,
         axesStyles,
         xDomain,
         [yDomain],
-        chartDim,
+        emptySmScales,
         1,
         false,
         customFotmatter,
@@ -1741,13 +1744,13 @@ describe('Axis computational utils', () => {
           leftMargin: 0,
         },
         LIGHT_THEME,
-        0,
+        NO_ROTATION,
         axisSpecs,
         axisDims,
         axesStyles,
         xDomain,
         [yDomain],
-        chartDim,
+        emptySmScales,
         1,
         false,
         customFotmatter,
@@ -1757,6 +1760,151 @@ describe('Axis computational utils', () => {
       expect(axisTicksPosition.find(({ axis: { id } }) => id === spec.id)!.ticks.map(({ label }) => label)).toEqual(
         expected,
       );
+    });
+  });
+
+  describe('Small multiples', () => {
+    const axisStyles = axisTitleStyles(10, 8);
+    const cumTopSum = 10;
+    const cumBottomSum = 10;
+    const cumLeftSum = 10;
+    const cumRightSum = 10;
+    const smScales = getSmScales(['a'], [0]);
+
+    describe.each(['test', ''])('Axes title positions - title is "%s"', (title) => {
+      test('should compute left axis position', () => {
+        const leftAxisPosition = getAxisPosition(
+          chartDim,
+          LIGHT_THEME.chartMargins,
+          axisStyles,
+          { ...verticalAxisSpec, title },
+          axis1Dims,
+          smScales,
+          cumTopSum,
+          cumBottomSum,
+          cumLeftSum,
+          cumRightSum,
+          10,
+          0,
+          true,
+        );
+
+        const expectedLeftAxisPosition = {
+          dimensions: {
+            height: 100,
+            width: title ? 56 : 36,
+            left: 110,
+            top: 0,
+          },
+          topIncrement: 0,
+          bottomIncrement: 0,
+          leftIncrement: 0,
+          rightIncrement: title ? 66 : 46,
+        };
+
+        expect(leftAxisPosition).toEqual(expectedLeftAxisPosition);
+      });
+
+      test('should compute right axis position', () => {
+        verticalAxisSpec.position = Position.Right;
+        const rightAxisPosition = getAxisPosition(
+          chartDim,
+          LIGHT_THEME.chartMargins,
+          axisStyles,
+          { ...verticalAxisSpec, title },
+          axis1Dims,
+          smScales,
+          cumTopSum,
+          cumBottomSum,
+          cumLeftSum,
+          cumRightSum,
+          10,
+          0,
+          true,
+        );
+
+        const expectedRightAxisPosition = {
+          dimensions: {
+            height: 100,
+            width: title ? 56 : 36,
+            left: 110,
+            top: 0,
+          },
+          topIncrement: 0,
+          bottomIncrement: 0,
+          leftIncrement: 0,
+          rightIncrement: title ? 66 : 46,
+        };
+
+        expect(rightAxisPosition).toEqual(expectedRightAxisPosition);
+      });
+
+      test('should compute top axis position', () => {
+        horizontalAxisSpec.position = Position.Top;
+        const topAxisPosition = getAxisPosition(
+          chartDim,
+          LIGHT_THEME.chartMargins,
+          axisStyles,
+          { ...horizontalAxisSpec, title },
+          axis1Dims,
+          smScales,
+          cumTopSum,
+          cumBottomSum,
+          cumLeftSum,
+          cumRightSum,
+          10,
+          0,
+          true,
+        );
+
+        const expectedTopAxisPosition = {
+          dimensions: {
+            height: title ? 56 : 36,
+            width: 100,
+            left: 0,
+            top: 20,
+          },
+          topIncrement: title ? 66 : 46,
+          bottomIncrement: 0,
+          leftIncrement: 0,
+          rightIncrement: 0,
+        };
+
+        expect(topAxisPosition).toEqual(expectedTopAxisPosition);
+      });
+
+      test('should compute bottom axis position', () => {
+        horizontalAxisSpec.position = Position.Bottom;
+        const bottomAxisPosition = getAxisPosition(
+          chartDim,
+          LIGHT_THEME.chartMargins,
+          axisStyles,
+          { ...horizontalAxisSpec, title },
+          axis1Dims,
+          smScales,
+          cumTopSum,
+          cumBottomSum,
+          cumLeftSum,
+          cumRightSum,
+          10,
+          0,
+          true,
+        );
+
+        const expectedBottomAxisPosition = {
+          dimensions: {
+            height: title ? 56 : 36,
+            width: 100,
+            left: 0,
+            top: 110,
+          },
+          topIncrement: 0,
+          bottomIncrement: title ? 66 : 46,
+          leftIncrement: 0,
+          rightIncrement: 0,
+        };
+        expect(bottomAxisPosition).toEqual(expectedBottomAxisPosition);
+      });
     });
   });
 });