+++ src/chart_types/xy_chart/utils/panel.ts
@@ -23,3 +23,7 @@ import { SmallMultipleScales } from '../state/selectors/compute_small_multiple_s
 export function getPanelSize({ horizontal, vertical }: SmallMultipleScales): Size {
   return { width: horizontal.bandwidth, height: vertical.bandwidth };
 }
+
+/** @internal */
+export const hasSMDomain = ({ domain }: SmallMultipleScales['horizontal'] | SmallMultipleScales['vertical']) =>
+  domain.length > 0 && domain[0] !== undefined;